Move components to js folder
