+++ Task-10_Browser-game/src/js/controls.js
