+++ Task-10_Browser-game/src/js/mapComponent.js
