+++ Task-10_Browser-game/src/js/menuComponent.js
