+++ Task-10_Browser-game/src/js/render.js
