+++ Task-10_Browser-game/src/js/utils.js
